Tests in eingene Klassen auslagern
